[miele] Made constants and handler packages internal
