+++ addons/binding/org.openhab.binding.miele/META-INF/MANIFEST.MF
@@ -7,9 +7,6 @@ Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Bundle-SymbolicName: org.openhab.binding.miele;singleton:=true
 Bundle-Vendor: openHAB
 Bundle-Version: 2.4.0.qualifier
-Export-Package: 
- org.openhab.binding.miele,
- org.openhab.binding.miele.handler
 Import-Package: 
  com.google.common.collect,
  com.google.gson,
@@ -25,8 +22,6 @@ Import-Package:
  org.eclipse.smarthome.core.thing.binding.builder,
  org.eclipse.smarthome.core.thing.type,
  org.eclipse.smarthome.core.types,
- org.openhab.binding.miele,
- org.openhab.binding.miele.handler,
  org.osgi.framework,
  org.slf4j
 Service-Component: OSGI-INF/*.xml
