--- addons/binding/org.openhab.binding.miele/src/main/java/org/openhab/binding/miele/handler/HoodChannelSelector.java
@@ -1,138 +1,0 @@
-/**
- * Copyright (c) 2010-2018 by the respective copyright holders.
- *
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
- */
-package org.openhab.binding.miele.handler;
-
-import java.lang.reflect.Method;
-import java.util.Map.Entry;
-
-import org.eclipse.smarthome.core.library.types.DecimalType;
-import org.eclipse.smarthome.core.library.types.OnOffType;
-import org.eclipse.smarthome.core.library.types.StringType;
-import org.eclipse.smarthome.core.types.State;
-import org.eclipse.smarthome.core.types.Type;
-import org.eclipse.smarthome.core.types.UnDefType;
-import org.openhab.binding.miele.handler.MieleBridgeHandler.DeviceMetaData;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import com.google.gson.JsonElement;
-
-/**
- * The {@link ApplianceChannelSelector} for ventilation hoods
- *
- * @author Karel Goderis - Initial contribution
- */
-public enum HoodChannelSelector implements ApplianceChannelSelector {
-
-    PRODUCT_TYPE("productTypeId", "productType", StringType.class, true),
-    DEVICE_TYPE("mieleDeviceType", "deviceType", StringType.class, true),
-    BRAND_ID("brandId", "brandId", StringType.class, true),
-    COMPANY_ID("companyId", "companyId", StringType.class, true),
-    STATE("state", "state", StringType.class, false),
-    VENTILATION("ventilationPower", "ventilation", DecimalType.class, false),
-    LIGHT("lightingStatus", "light", OnOffType.class, false) {
-        @Override
-
-        public State getState(String s, DeviceMetaData dmd) {
-            if ("true".equals(s)) {
-                return getState("ON");
-            }
-
-            if ("false".equals(s)) {
-                return getState("OFF");
-            }
-
-            return UnDefType.UNDEF;
-        }
-    },
-    STOP(null, "stop", OnOffType.class, false);
-
-    private final Logger logger = LoggerFactory.getLogger(HoodChannelSelector.class);
-
-    private final String mieleID;
-    private final String channelID;
-    private final Class<? extends Type> typeClass;
-    private final boolean isProperty;
-
-    HoodChannelSelector(String propertyID, String channelID, Class<? extends Type> typeClass,
-            boolean isProperty) {
-        this.mieleID = propertyID;
-        this.channelID = channelID;
-        this.typeClass = typeClass;
-        this.isProperty = isProperty;
-    }
-
-    @Override
-    public String toString() {
-        return mieleID;
-    }
-
-    @Override
-    public String getMieleID() {
-        return mieleID;
-    }
-
-    @Override
-    public String getChannelID() {
-        return channelID;
-    }
-
-    @Override
-    public Class<? extends Type> getTypeClass() {
-        return typeClass;
-    }
-
-    @Override
-    public boolean isProperty() {
-        return isProperty;
-    }
-
-    @Override
-    public State getState(String s, DeviceMetaData dmd) {
-        if (dmd != null) {
-            String localizedValue = getMieleEnum(s, dmd);
-            if (localizedValue == null) {
-                localizedValue = dmd.LocalizedValue;
-            }
-            if (localizedValue == null) {
-                localizedValue = s;
-            }
-
-            return getState(localizedValue);
-        } else {
-            return getState(s);
-        }
-    }
-
-    public State getState(String s) {
-        try {
-            Method valueOf = typeClass.getMethod("valueOf", String.class);
-            State state = (State) valueOf.invoke(typeClass, s);
-            if (state != null) {
-                return state;
-            }
-        } catch (Exception e) {
-            logger.error("An exception occurred while converting '{}' into a State", s);
-        }
-
-        return null;
-    }
-
-    public String getMieleEnum(String s, DeviceMetaData dmd) {
-        if (dmd.MieleEnum != null) {
-            for (Entry<String, JsonElement> enumEntry : dmd.MieleEnum.entrySet()) {
-                if (enumEntry.getValue().getAsString().trim().equals(s.trim())) {
-                    return enumEntry.getKey();
-                }
-            }
-        }
-
-        return null;
-    }
-}
