+++ addons/binding/org.openhab.binding.miele/src/main/java/org/openhab/binding/miele/internal/discovery/MieleMDNSDiscoveryParticipant.java
@@ -23,7 +23,7 @@ import org.eclipse.smarthome.config.discovery.DiscoveryResultBuilder;
 import org.eclipse.smarthome.config.discovery.mdns.MDNSDiscoveryParticipant;
 import org.eclipse.smarthome.core.thing.ThingTypeUID;
 import org.eclipse.smarthome.core.thing.ThingUID;
-import org.openhab.binding.miele.MieleBindingConstants;
+import org.openhab.binding.miele.internal.MieleBindingConstants;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
