+++ addons/binding/org.openhab.binding.miele/src/main/java/org/openhab/binding/miele/internal/handler/DishWasherHandler.java
@@ -6,32 +6,33 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.miele.handler;
+package org.openhab.binding.miele.internal.handler;
 
-import static org.openhab.binding.miele.MieleBindingConstants.APPLIANCE_ID;
+import static org.openhab.binding.miele.internal.MieleBindingConstants.APPLIANCE_ID;
 
 import org.eclipse.smarthome.core.library.types.OnOffType;
 import org.eclipse.smarthome.core.thing.ChannelUID;
 import org.eclipse.smarthome.core.thing.Thing;
 import org.eclipse.smarthome.core.types.Command;
+import org.eclipse.smarthome.core.types.RefreshType;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.google.gson.JsonElement;
 
 /**
- * The {@link HoodHandler} is responsible for handling commands,
+ * The {@link DishWasherHandler} is responsible for handling commands,
  * which are sent to one of the channels
  *
  * @author Karel Goderis - Initial contribution
  * @author Kai Kreuzer - fixed handling of REFRESH commands
  */
-public class HoodHandler extends MieleApplianceHandler<HoodChannelSelector> {
+public class DishWasherHandler extends MieleApplianceHandler<DishwasherChannelSelector> {
 
-    private final Logger logger = LoggerFactory.getLogger(HoodHandler.class);
+    private final Logger logger = LoggerFactory.getLogger(DishWasherHandler.class);
 
-    public HoodHandler(Thing thing) {
-        super(thing, HoodChannelSelector.class, "Hood");
+    public DishWasherHandler(Thing thing) {
+        super(thing, DishwasherChannelSelector.class, "Dishwasher");
     }
 
     @Override
@@ -42,29 +43,25 @@ public class HoodHandler extends MieleApplianceHandler<HoodChannelSelector> {
         String channelID = channelUID.getId();
         String uid = (String) getThing().getConfiguration().getProperties().get(APPLIANCE_ID);
 
-        HoodChannelSelector selector = (HoodChannelSelector) getValueSelectorFromChannelID(channelID);
+        DishwasherChannelSelector selector = (DishwasherChannelSelector) getValueSelectorFromChannelID(channelID);
         JsonElement result = null;
 
         try {
             if (selector != null) {
                 switch (selector) {
-                    case LIGHT: {
+                    case SWITCH: {
                         if (command.equals(OnOffType.ON)) {
-                            result = bridgeHandler.invokeOperation(uid, modelID, "startLighting");
+                            result = bridgeHandler.invokeOperation(uid, modelID, "start");
                         } else if (command.equals(OnOffType.OFF)) {
-                            result = bridgeHandler.invokeOperation(uid, modelID, "stopLighting");
-                        }
-                        break;
-                    }
-                    case STOP: {
-                        if (command.equals(OnOffType.ON)) {
                             result = bridgeHandler.invokeOperation(uid, modelID, "stop");
                         }
                         break;
                     }
                     default: {
-                        logger.debug("{} is a read-only channel that does not accept commands",
-                                selector.getChannelID());
+                        if (!(command instanceof RefreshType)) {
+                            logger.debug("{} is a read-only channel that does not accept commands",
+                                    selector.getChannelID());
+                        }
                     }
                 }
             }
