+++ addons/binding/org.openhab.binding.miele/src/main/java/org/openhab/binding/miele/internal/handler/FridgeChannelSelector.java
@@ -6,15 +6,11 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.miele.handler;
+package org.openhab.binding.miele.internal.handler;
 
 import java.lang.reflect.Method;
-import java.text.SimpleDateFormat;
-import java.util.Date;
 import java.util.Map.Entry;
-import java.util.TimeZone;
 
-import org.eclipse.smarthome.core.library.types.DateTimeType;
 import org.eclipse.smarthome.core.library.types.DecimalType;
 import org.eclipse.smarthome.core.library.types.OnOffType;
 import org.eclipse.smarthome.core.library.types.OpenClosedType;
@@ -22,102 +18,32 @@ import org.eclipse.smarthome.core.library.types.StringType;
 import org.eclipse.smarthome.core.types.State;
 import org.eclipse.smarthome.core.types.Type;
 import org.eclipse.smarthome.core.types.UnDefType;
-import org.openhab.binding.miele.handler.MieleBridgeHandler.DeviceMetaData;
+import org.openhab.binding.miele.internal.handler.MieleBridgeHandler.DeviceMetaData;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.google.gson.JsonElement;
 
 /**
- * The {@link ApplianceChannelSelector} for ovens
+ * The {@link ApplianceChannelSelector} for fridges
  *
  * @author Karel Goderis - Initial contribution
- * @author Kai Kreuzer - Changed START_TIME to DateTimeType
  */
-public enum OvenChannelSelector implements ApplianceChannelSelector {
+public enum FridgeChannelSelector implements ApplianceChannelSelector {
 
     PRODUCT_TYPE("productTypeId", "productType", StringType.class, true),
     DEVICE_TYPE("mieleDeviceType", "deviceType", StringType.class, true),
     BRAND_ID("brandId", "brandId", StringType.class, true),
     COMPANY_ID("companyId", "companyId", StringType.class, true),
     STATE("state", "state", StringType.class, false),
-    PROGRAMID("programId", "program", StringType.class, false),
-    PROGRAMPHASE("phase", "phase", StringType.class, false),
-    START_TIME("startTime", "start", DateTimeType.class, false) {
-        @Override
-        public State getState(String s, DeviceMetaData dmd) {
-            Date date = new Date();
-            SimpleDateFormat DATE_FORMATTER = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss");
-            DATE_FORMATTER.setTimeZone(TimeZone.getTimeZone("GMT+0"));
-            try {
-                date.setTime(Long.valueOf(s) * 60000);
-            } catch (Exception e) {
-                date.setTime(0);
-            }
-            return getState(DATE_FORMATTER.format(date));
-        }
-    },
-    DURATION("duration", "duration", DateTimeType.class, false) {
-        @Override
-        public State getState(String s, DeviceMetaData dmd) {
-            Date date = new Date();
-            SimpleDateFormat DATE_FORMATTER = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss");
-            DATE_FORMATTER.setTimeZone(TimeZone.getTimeZone("GMT+0"));
-            try {
-                date.setTime(Long.valueOf(s) * 60000);
-            } catch (Exception e) {
-                date.setTime(0);
-            }
-            return getState(DATE_FORMATTER.format(date));
-        }
-    },
-    ELAPSED_TIME("elapsedTime", "elapsed", DateTimeType.class, false) {
-        @Override
-        public State getState(String s, DeviceMetaData dmd) {
-            Date date = new Date();
-            SimpleDateFormat DATE_FORMATTER = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss");
-            DATE_FORMATTER.setTimeZone(TimeZone.getTimeZone("GMT+0"));
-            try {
-                date.setTime(Long.valueOf(s) * 60000);
-            } catch (Exception e) {
-                date.setTime(0);
-            }
-            return getState(DATE_FORMATTER.format(date));
-        }
-    },
-    FINISH_TIME("finishTime", "finish", DateTimeType.class, false) {
-        @Override
-        public State getState(String s, DeviceMetaData dmd) {
-            Date date = new Date();
-            SimpleDateFormat DATE_FORMATTER = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss");
-            DATE_FORMATTER.setTimeZone(TimeZone.getTimeZone("GMT+0"));
-            try {
-                date.setTime(Long.valueOf(s) * 60000);
-            } catch (Exception e) {
-                date.setTime(0);
-            }
-            return getState(DATE_FORMATTER.format(date));
-        }
-    },
-    TARGET_TEMP("targetTemperature", "target", DecimalType.class, false) {
+    SUPERCOOL(null, "supercool", OnOffType.class, false),
+    FRIDGECURRENTTEMP("currentTemperature", "current", DecimalType.class, false) {
         @Override
         public State getState(String s, DeviceMetaData dmd) {
             return getState(s);
         }
     },
-    MEASURED_TEMP("measuredTemperature", "measured", DecimalType.class, false) {
-        @Override
-        public State getState(String s, DeviceMetaData dmd) {
-            return getState(s);
-        }
-    },
-    DEVICE_TEMP_ONE("deviceTemperature1", "temp1", DecimalType.class, false) {
-        @Override
-        public State getState(String s, DeviceMetaData dmd) {
-            return getState(s);
-        }
-    },
-    DEVICE_TEMP_TWO("deviceTemperature2", "temp2", DecimalType.class, false) {
+    FRIDGETARGETTEMP("targetTemperature", "target", DecimalType.class, false) {
         @Override
         public State getState(String s, DeviceMetaData dmd) {
             return getState(s);
@@ -138,17 +64,16 @@ public enum OvenChannelSelector implements ApplianceChannelSelector {
             return UnDefType.UNDEF;
         }
     },
-    STOP(null, "stop", OnOffType.class, false),
-    SWITCH(null, "switch", OnOffType.class, false);
+    START(null, "start", OnOffType.class, false);
 
-    private final Logger logger = LoggerFactory.getLogger(OvenChannelSelector.class);
+    private final Logger logger = LoggerFactory.getLogger(FridgeChannelSelector.class);
 
     private final String mieleID;
     private final String channelID;
     private final Class<? extends Type> typeClass;
     private final boolean isProperty;
 
-    OvenChannelSelector(String propertyID, String channelID, Class<? extends Type> typeClass,
+    FridgeChannelSelector(String propertyID, String channelID, Class<? extends Type> typeClass,
             boolean isProperty) {
         this.mieleID = propertyID;
         this.channelID = channelID;
