+++ addons/binding/org.openhab.binding.miele/src/main/java/org/openhab/binding/miele/internal/handler/FridgeFreezerHandler.java
@@ -6,33 +6,32 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.miele.handler;
+package org.openhab.binding.miele.internal.handler;
 
-import static org.openhab.binding.miele.MieleBindingConstants.APPLIANCE_ID;
+import static org.openhab.binding.miele.internal.MieleBindingConstants.APPLIANCE_ID;
 
 import org.eclipse.smarthome.core.library.types.OnOffType;
 import org.eclipse.smarthome.core.thing.ChannelUID;
 import org.eclipse.smarthome.core.thing.Thing;
 import org.eclipse.smarthome.core.types.Command;
-import org.eclipse.smarthome.core.types.RefreshType;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.google.gson.JsonElement;
 
 /**
- * The {@link DishWasherHandler} is responsible for handling commands,
+ * The {@link FridgeFreezerHandler} is responsible for handling commands,
  * which are sent to one of the channels
  *
  * @author Karel Goderis - Initial contribution
  * @author Kai Kreuzer - fixed handling of REFRESH commands
  */
-public class DishWasherHandler extends MieleApplianceHandler<DishwasherChannelSelector> {
+public class FridgeFreezerHandler extends MieleApplianceHandler<FridgeFreezerChannelSelector> {
 
-    private final Logger logger = LoggerFactory.getLogger(DishWasherHandler.class);
+    private final Logger logger = LoggerFactory.getLogger(FridgeFreezerHandler.class);
 
-    public DishWasherHandler(Thing thing) {
-        super(thing, DishwasherChannelSelector.class, "Dishwasher");
+    public FridgeFreezerHandler(Thing thing) {
+        super(thing, FridgeFreezerChannelSelector.class, "FridgeFreezer");
     }
 
     @Override
@@ -43,25 +42,31 @@ public class DishWasherHandler extends MieleApplianceHandler<DishwasherChannelSe
         String channelID = channelUID.getId();
         String uid = (String) getThing().getConfiguration().getProperties().get(APPLIANCE_ID);
 
-        DishwasherChannelSelector selector = (DishwasherChannelSelector) getValueSelectorFromChannelID(channelID);
+        FridgeFreezerChannelSelector selector = (FridgeFreezerChannelSelector) getValueSelectorFromChannelID(channelID);
         JsonElement result = null;
 
         try {
             if (selector != null) {
                 switch (selector) {
-                    case SWITCH: {
+                    case SUPERCOOL: {
                         if (command.equals(OnOffType.ON)) {
-                            result = bridgeHandler.invokeOperation(uid, modelID, "start");
+                            result = bridgeHandler.invokeOperation(uid, modelID, "startSuperCooling");
                         } else if (command.equals(OnOffType.OFF)) {
-                            result = bridgeHandler.invokeOperation(uid, modelID, "stop");
+                            result = bridgeHandler.invokeOperation(uid, modelID, "stopSuperCooling");
+                        }
+                        break;
+                    }
+                    case SUPERFREEZE: {
+                        if (command.equals(OnOffType.ON)) {
+                            result = bridgeHandler.invokeOperation(uid, modelID, "startSuperFreezing");
+                        } else if (command.equals(OnOffType.OFF)) {
+                            result = bridgeHandler.invokeOperation(uid, modelID, "stopSuperFreezing");
                         }
                         break;
                     }
                     default: {
-                        if (!(command instanceof RefreshType)) {
-                            logger.debug("{} is a read-only channel that does not accept commands",
-                                    selector.getChannelID());
-                        }
+                        logger.debug("{} is a read-only channel that does not accept commands",
+                                selector.getChannelID());
                     }
                 }
             }
