+++ addons/binding/org.openhab.binding.miele/src/main/java/org/openhab/binding/miele/internal/handler/OvenChannelSelector.java
@@ -6,11 +6,15 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.miele.handler;
+package org.openhab.binding.miele.internal.handler;
 
 import java.lang.reflect.Method;
+import java.text.SimpleDateFormat;
+import java.util.Date;
 import java.util.Map.Entry;
+import java.util.TimeZone;
 
+import org.eclipse.smarthome.core.library.types.DateTimeType;
 import org.eclipse.smarthome.core.library.types.DecimalType;
 import org.eclipse.smarthome.core.library.types.OnOffType;
 import org.eclipse.smarthome.core.library.types.OpenClosedType;
@@ -18,48 +22,102 @@ import org.eclipse.smarthome.core.library.types.StringType;
 import org.eclipse.smarthome.core.types.State;
 import org.eclipse.smarthome.core.types.Type;
 import org.eclipse.smarthome.core.types.UnDefType;
-import org.openhab.binding.miele.handler.MieleBridgeHandler.DeviceMetaData;
+import org.openhab.binding.miele.internal.handler.MieleBridgeHandler.DeviceMetaData;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.google.gson.JsonElement;
 
 /**
- * The {@link ApplianceChannelSelector} for fridges with
- * a freezer compartment
+ * The {@link ApplianceChannelSelector} for ovens
  *
  * @author Karel Goderis - Initial contribution
+ * @author Kai Kreuzer - Changed START_TIME to DateTimeType
  */
-public enum FridgeFreezerChannelSelector implements ApplianceChannelSelector {
+public enum OvenChannelSelector implements ApplianceChannelSelector {
 
     PRODUCT_TYPE("productTypeId", "productType", StringType.class, true),
     DEVICE_TYPE("mieleDeviceType", "deviceType", StringType.class, true),
     BRAND_ID("brandId", "brandId", StringType.class, true),
     COMPANY_ID("companyId", "companyId", StringType.class, true),
     STATE("state", "state", StringType.class, false),
-    FREEZERSTATE("freezerState", "freezerstate", StringType.class, false),
-    FRIDGESTATE("fridgeState", "fridgestate", StringType.class, false),
-    SUPERCOOL(null, "supercool", OnOffType.class, false),
-    SUPERFREEZE(null, "superfreeze", OnOffType.class, false),
-    FREEZERCURRENTTEMP("freezerCurrentTemperature", "freezercurrent", DecimalType.class, false) {
+    PROGRAMID("programId", "program", StringType.class, false),
+    PROGRAMPHASE("phase", "phase", StringType.class, false),
+    START_TIME("startTime", "start", DateTimeType.class, false) {
+        @Override
+        public State getState(String s, DeviceMetaData dmd) {
+            Date date = new Date();
+            SimpleDateFormat DATE_FORMATTER = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss");
+            DATE_FORMATTER.setTimeZone(TimeZone.getTimeZone("GMT+0"));
+            try {
+                date.setTime(Long.valueOf(s) * 60000);
+            } catch (Exception e) {
+                date.setTime(0);
+            }
+            return getState(DATE_FORMATTER.format(date));
+        }
+    },
+    DURATION("duration", "duration", DateTimeType.class, false) {
+        @Override
+        public State getState(String s, DeviceMetaData dmd) {
+            Date date = new Date();
+            SimpleDateFormat DATE_FORMATTER = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss");
+            DATE_FORMATTER.setTimeZone(TimeZone.getTimeZone("GMT+0"));
+            try {
+                date.setTime(Long.valueOf(s) * 60000);
+            } catch (Exception e) {
+                date.setTime(0);
+            }
+            return getState(DATE_FORMATTER.format(date));
+        }
+    },
+    ELAPSED_TIME("elapsedTime", "elapsed", DateTimeType.class, false) {
+        @Override
+        public State getState(String s, DeviceMetaData dmd) {
+            Date date = new Date();
+            SimpleDateFormat DATE_FORMATTER = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss");
+            DATE_FORMATTER.setTimeZone(TimeZone.getTimeZone("GMT+0"));
+            try {
+                date.setTime(Long.valueOf(s) * 60000);
+            } catch (Exception e) {
+                date.setTime(0);
+            }
+            return getState(DATE_FORMATTER.format(date));
+        }
+    },
+    FINISH_TIME("finishTime", "finish", DateTimeType.class, false) {
+        @Override
+        public State getState(String s, DeviceMetaData dmd) {
+            Date date = new Date();
+            SimpleDateFormat DATE_FORMATTER = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss");
+            DATE_FORMATTER.setTimeZone(TimeZone.getTimeZone("GMT+0"));
+            try {
+                date.setTime(Long.valueOf(s) * 60000);
+            } catch (Exception e) {
+                date.setTime(0);
+            }
+            return getState(DATE_FORMATTER.format(date));
+        }
+    },
+    TARGET_TEMP("targetTemperature", "target", DecimalType.class, false) {
         @Override
         public State getState(String s, DeviceMetaData dmd) {
             return getState(s);
         }
     },
-    FREEZERTARGETTEMP("freezerTargetTemperature", "freezertarget", DecimalType.class, false) {
+    MEASURED_TEMP("measuredTemperature", "measured", DecimalType.class, false) {
         @Override
         public State getState(String s, DeviceMetaData dmd) {
             return getState(s);
         }
     },
-    FRIDGECURRENTTEMP("fridgeCurrentTemperature", "fridgecurrent", DecimalType.class, false) {
+    DEVICE_TEMP_ONE("deviceTemperature1", "temp1", DecimalType.class, false) {
         @Override
         public State getState(String s, DeviceMetaData dmd) {
             return getState(s);
         }
     },
-    FRIDGETARGETTEMP("fridgeTargetTemperature", "fridgetarget", DecimalType.class, false) {
+    DEVICE_TEMP_TWO("deviceTemperature2", "temp2", DecimalType.class, false) {
         @Override
         public State getState(String s, DeviceMetaData dmd) {
             return getState(s);
@@ -80,16 +138,17 @@ public enum FridgeFreezerChannelSelector implements ApplianceChannelSelector {
             return UnDefType.UNDEF;
         }
     },
-    START(null, "start", OnOffType.class, false);
+    STOP(null, "stop", OnOffType.class, false),
+    SWITCH(null, "switch", OnOffType.class, false);
 
-    private final Logger logger = LoggerFactory.getLogger(FridgeFreezerChannelSelector.class);
+    private final Logger logger = LoggerFactory.getLogger(OvenChannelSelector.class);
 
     private final String mieleID;
     private final String channelID;
     private final Class<? extends Type> typeClass;
     private final boolean isProperty;
 
-    FridgeFreezerChannelSelector(String propertyID, String channelID, Class<? extends Type> typeClass,
+    OvenChannelSelector(String propertyID, String channelID, Class<? extends Type> typeClass,
             boolean isProperty) {
         this.mieleID = propertyID;
         this.channelID = channelID;
