+++ addons/binding/org.openhab.binding.miele/src/main/java/org/openhab/binding/miele/internal/handler/WashingMachineHandler.java
@@ -6,9 +6,9 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.miele.handler;
+package org.openhab.binding.miele.internal.handler;
 
-import static org.openhab.binding.miele.MieleBindingConstants.APPLIANCE_ID;
+import static org.openhab.binding.miele.internal.MieleBindingConstants.APPLIANCE_ID;
 
 import org.eclipse.smarthome.core.library.types.OnOffType;
 import org.eclipse.smarthome.core.thing.ChannelUID;
@@ -21,18 +21,18 @@ import org.slf4j.LoggerFactory;
 import com.google.gson.JsonElement;
 
 /**
- * The {@link TumbleDryerHandler} is responsible for handling commands,
+ * The {@link WashingMachineHandler} is responsible for handling commands,
  * which are sent to one of the channels
  *
  * @author Karel Goderis - Initial contribution
  * @author Kai Kreuzer - fixed handling of REFRESH commands
  */
-public class TumbleDryerHandler extends MieleApplianceHandler<TumbleDryerChannelSelector> {
+public class WashingMachineHandler extends MieleApplianceHandler<WashingMachineChannelSelector> {
 
-    private final Logger logger = LoggerFactory.getLogger(TumbleDryerHandler.class);
+    private final Logger logger = LoggerFactory.getLogger(WashingMachineHandler.class);
 
-    public TumbleDryerHandler(Thing thing) {
-        super(thing, TumbleDryerChannelSelector.class, "TumbleDryer");
+    public WashingMachineHandler(Thing thing) {
+        super(thing, WashingMachineChannelSelector.class, "WashingMachine");
     }
 
     @Override
@@ -43,7 +43,8 @@ public class TumbleDryerHandler extends MieleApplianceHandler<TumbleDryerChannel
         String channelID = channelUID.getId();
         String uid = (String) getThing().getConfiguration().getProperties().get(APPLIANCE_ID);
 
-        TumbleDryerChannelSelector selector = (TumbleDryerChannelSelector) getValueSelectorFromChannelID(channelID);
+        WashingMachineChannelSelector selector = (WashingMachineChannelSelector) getValueSelectorFromChannelID(
+                channelID);
         JsonElement result = null;
 
         try {
